:heavy_plus_sign: :hammer: :man_juggling: :gear: Add Skills Section. Introduce multiple @mixins in _helpers.scss
Add the @mixin flex(arguments...) with arguments passed to put the elements in a position using display:flex or with Flexbox CSS.
Add the @mixin pseudo-elements {...} with @content used to setting up the boilerplate for work with ::after & ::before pseudo-elements and their positions relative|absolute depend of the element parent|child, respectively.
Change the flex-center mixin, to class. This would be rollback to the mixin version, but for now the .class implementation is more practical.
Introduce the global padding body variable. It's to make uniform the padding throughout the page|site.